Remove some tchap classes
