--- vector/src/main/java/fr/gouv/tchap/features/home/TchapHomeActivity.kt
@@ -1,427 +1,0 @@
-/*
- * Copyright 2019 New Vector Ltd
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package fr.gouv.tchap.features.home
-
-import android.content.Context
-import android.content.Intent
-import android.net.Uri
-import android.os.Bundle
-import android.os.Parcelable
-import android.view.MenuItem
-import androidx.appcompat.app.AlertDialog
-import androidx.appcompat.widget.Toolbar
-import androidx.core.content.ContextCompat
-import androidx.core.view.GravityCompat
-import androidx.core.view.isVisible
-import androidx.drawerlayout.widget.DrawerLayout
-import com.airbnb.mvrx.MvRx
-import com.airbnb.mvrx.viewModel
-import im.vector.app.R
-import im.vector.app.core.di.ActiveSessionHolder
-import im.vector.app.core.di.ScreenComponent
-import im.vector.app.core.extensions.exhaustive
-import im.vector.app.core.extensions.hideKeyboard
-import im.vector.app.core.extensions.replaceFragment
-import im.vector.app.core.platform.ToolbarConfigurable
-import im.vector.app.core.platform.VectorBaseActivity
-import im.vector.app.core.pushers.PushersManager
-import im.vector.app.databinding.ActivityHomeBinding
-import im.vector.app.features.disclaimer.showDisclaimerDialog
-import im.vector.app.features.home.AvatarRenderer
-import im.vector.app.features.home.HomeActivitySharedAction
-import im.vector.app.features.home.HomeActivityViewActions
-import im.vector.app.features.home.HomeActivityViewEvents
-import im.vector.app.features.home.HomeActivityViewState
-import im.vector.app.features.home.HomeDrawerFragment
-import im.vector.app.features.home.HomeSharedActionViewModel
-import im.vector.app.features.home.LoadingFragment
-import im.vector.app.features.home.ShortcutsHandler
-import im.vector.app.features.home.UnknownDeviceDetectorSharedViewModel
-import im.vector.app.features.home.UnknownDevicesState
-import im.vector.app.features.invite.InviteUsersToRoomViewModel
-import im.vector.app.features.matrixto.MatrixToBottomSheet
-import im.vector.app.features.notifications.NotificationDrawerManager
-import im.vector.app.features.permalink.NavigationInterceptor
-import im.vector.app.features.permalink.PermalinkHandler
-import im.vector.app.features.popup.DefaultVectorAlert
-import im.vector.app.features.popup.PopupAlertManager
-import im.vector.app.features.popup.VerificationVectorAlert
-import im.vector.app.features.rageshake.VectorUncaughtExceptionHandler
-import im.vector.app.features.settings.VectorPreferences
-import im.vector.app.features.settings.VectorSettingsActivity
-import im.vector.app.features.themes.ThemeUtils
-import im.vector.app.features.userdirectory.UserListViewModel
-import im.vector.app.features.userdirectory.UserListViewState
-import im.vector.app.features.workers.signout.ServerBackupStatusViewModel
-import im.vector.app.features.workers.signout.ServerBackupStatusViewState
-import im.vector.app.push.fcm.FcmHelper
-import io.reactivex.android.schedulers.AndroidSchedulers
-import kotlinx.parcelize.Parcelize
-import org.matrix.android.sdk.api.session.InitialSyncProgressService
-import org.matrix.android.sdk.api.session.permalinks.PermalinkService
-import org.matrix.android.sdk.api.util.MatrixItem
-import timber.log.Timber
-import javax.inject.Inject
-
-@Parcelize
-data class HomeActivityArgs(
-        val clearNotification: Boolean,
-        val accountCreation: Boolean
-) : Parcelable
-
-class TchapHomeActivity :
-        VectorBaseActivity<ActivityHomeBinding>(),
-        ToolbarConfigurable,
-        UnknownDeviceDetectorSharedViewModel.Factory,
-        ServerBackupStatusViewModel.Factory,
-        UserListViewModel.Factory,
-        NavigationInterceptor {
-
-    private lateinit var sharedActionViewModel: HomeSharedActionViewModel
-
-    private val homeActivityViewModel: TchapHomeActivityViewModel by viewModel()
-    @Inject lateinit var viewModelFactory: TchapHomeActivityViewModel.Factory
-
-    private val serverBackupStatusViewModel: ServerBackupStatusViewModel by viewModel()
-    @Inject lateinit var serverBackupviewModelFactory: ServerBackupStatusViewModel.Factory
-
-    @Inject lateinit var activeSessionHolder: ActiveSessionHolder
-    @Inject lateinit var vectorUncaughtExceptionHandler: VectorUncaughtExceptionHandler
-    @Inject lateinit var pushManager: PushersManager
-    @Inject lateinit var notificationDrawerManager: NotificationDrawerManager
-    @Inject lateinit var vectorPreferences: VectorPreferences
-    @Inject lateinit var popupAlertManager: PopupAlertManager
-    @Inject lateinit var shortcutsHandler: ShortcutsHandler
-    @Inject lateinit var unknownDeviceViewModelFactory: UnknownDeviceDetectorSharedViewModel.Factory
-    @Inject lateinit var permalinkHandler: PermalinkHandler
-    @Inject lateinit var avatarRenderer: AvatarRenderer
-    @Inject lateinit var userListViewModelFactory: UserListViewModel.Factory
-
-    private val drawerListener = object : DrawerLayout.SimpleDrawerListener() {
-        override fun onDrawerStateChanged(newState: Int) {
-            hideKeyboard()
-        }
-    }
-
-    override fun getBinding() = ActivityHomeBinding.inflate(layoutInflater)
-
-    override fun injectWith(injector: ScreenComponent) {
-        injector.inject(this)
-    }
-
-    override fun create(initialState: UnknownDevicesState): UnknownDeviceDetectorSharedViewModel {
-        return unknownDeviceViewModelFactory.create(initialState)
-    }
-
-    override fun create(initialState: ServerBackupStatusViewState): ServerBackupStatusViewModel {
-        return serverBackupviewModelFactory.create(initialState)
-    }
-
-    override fun create(initialState: UserListViewState) = userListViewModelFactory.create(initialState)
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        FcmHelper.ensureFcmTokenIsRetrieved(this, pushManager, vectorPreferences.areNotificationEnabledForDevice())
-        sharedActionViewModel = viewModelProvider.get(HomeSharedActionViewModel::class.java)
-        views.drawerLayout.addDrawerListener(drawerListener)
-        if (isFirstCreation()) {
-            replaceFragment(R.id.homeDetailFragmentContainer, LoadingFragment::class.java)
-            replaceFragment(R.id.homeDrawerFragmentContainer, HomeDrawerFragment::class.java)
-        }
-
-        sharedActionViewModel
-                .observe()
-                .subscribe { sharedAction ->
-                    when (sharedAction) {
-                        is HomeActivitySharedAction.OpenDrawer  -> views.drawerLayout.openDrawer(GravityCompat.START)
-                        is HomeActivitySharedAction.CloseDrawer -> views.drawerLayout.closeDrawer(GravityCompat.START)
-                        is HomeActivitySharedAction.OpenGroup   -> {
-                            views.drawerLayout.closeDrawer(GravityCompat.START)
-                            replaceFragment(R.id.homeDetailFragmentContainer, TchapHomeDetailFragment::class.java, allowStateLoss = true)
-                        }
-                    }.exhaustive
-                }
-                .disposeOnDestroy()
-
-        val args = intent.getParcelableExtra<HomeActivityArgs>(MvRx.KEY_ARG)
-
-        if (args?.clearNotification == true) {
-            notificationDrawerManager.clearAllEvents()
-        }
-
-        homeActivityViewModel.observeViewEvents {
-            when (it) {
-                is HomeActivityViewEvents.AskPasswordToInitCrossSigning -> handleAskPasswordToInitCrossSigning(it)
-                is HomeActivityViewEvents.OnNewSession                  -> handleOnNewSession(it)
-                HomeActivityViewEvents.PromptToEnableSessionPush -> handlePromptToEnablePush()
-                is HomeActivityViewEvents.OnCrossSignedInvalidated -> handleCrossSigningInvalidated(it)
-            }.exhaustive
-        }
-        homeActivityViewModel.subscribe(this) { renderState(it) }
-
-        shortcutsHandler.observeRoomsAndBuildShortcuts()
-                .disposeOnDestroy()
-
-        if (isFirstCreation()) {
-            handleIntent(intent)
-        }
-    }
-
-    private fun handleIntent(intent: Intent?) {
-        intent?.dataString?.let { deepLink ->
-            val resolvedLink = when {
-                deepLink.startsWith(PermalinkService.MATRIX_TO_URL_BASE) -> deepLink
-                deepLink.startsWith(MATRIX_TO_CUSTOM_SCHEME_URL_BASE)    -> {
-                    // This is a bit ugly, but for now just convert to matrix.to link for compatibility
-                    when {
-                        deepLink.startsWith(USER_LINK_PREFIX) -> deepLink.substring(USER_LINK_PREFIX.length)
-                        deepLink.startsWith(ROOM_LINK_PREFIX) -> deepLink.substring(ROOM_LINK_PREFIX.length)
-                        else                                  -> null
-                    }?.let {
-                        activeSessionHolder.getSafeActiveSession()?.permalinkService()?.createPermalink(it)
-                    }
-                }
-                else                                                     -> return@let
-            }
-
-            permalinkHandler.launch(
-                    context = this,
-                    deepLink = resolvedLink,
-                    navigationInterceptor = this,
-                    buildTask = true
-            )
-                    // .delay(500, TimeUnit.MILLISECONDS)
-                    .observeOn(AndroidSchedulers.mainThread())
-                    .subscribe { isHandled ->
-                        if (!isHandled) {
-                            AlertDialog.Builder(this)
-                                    .setTitle(R.string.dialog_title_error)
-                                    .setMessage(R.string.permalink_malformed)
-                                    .setPositiveButton(R.string.ok, null)
-                                    .show()
-                        }
-                    }
-                    .disposeOnDestroy()
-        }
-    }
-
-    private fun renderState(state: HomeActivityViewState) {
-        when (val status = state.initialSyncProgressServiceStatus) {
-            is InitialSyncProgressService.Status.Idle -> {
-                views.waitingView.root.isVisible = false
-            }
-            is InitialSyncProgressService.Status.Progressing -> {
-                Timber.v("${getString(status.statusText)} ${status.percentProgress}")
-                views.waitingView.root.setOnClickListener {
-                    // block interactions
-                }
-                views.waitingView.waitingHorizontalProgress.apply {
-                    isIndeterminate = false
-                    max = 100
-                    progress = status.percentProgress
-                    isVisible = true
-                }
-                views.waitingView.waitingStatusText.apply {
-                    text = getString(status.statusText)
-                    isVisible = true
-                }
-                views.waitingView.root.isVisible = true
-            }
-        }.exhaustive
-    }
-
-    private fun handleAskPasswordToInitCrossSigning(events: HomeActivityViewEvents.AskPasswordToInitCrossSigning) {
-        // We need to ask
-        promptSecurityEvent(
-                events.userItem,
-                R.string.upgrade_security,
-                R.string.security_prompt_text
-        ) {
-            it.navigator.upgradeSessionSecurity(it, true)
-        }
-    }
-
-    private fun handleCrossSigningInvalidated(event: HomeActivityViewEvents.OnCrossSignedInvalidated) {
-        // We need to ask
-        promptSecurityEvent(
-                event.userItem,
-                R.string.crosssigning_verify_this_session,
-                R.string.confirm_your_identity
-        ) {
-            it.navigator.waitSessionVerification(it)
-        }
-    }
-
-    private fun handleOnNewSession(event: HomeActivityViewEvents.OnNewSession) {
-        // We need to ask
-        promptSecurityEvent(
-                event.userItem,
-                R.string.crosssigning_verify_this_session,
-                R.string.confirm_your_identity
-        ) {
-            if (event.waitForIncomingRequest) {
-                it.navigator.waitSessionVerification(it)
-            } else {
-                it.navigator.requestSelfSessionVerification(it)
-            }
-        }
-    }
-
-    private fun handlePromptToEnablePush() {
-        popupAlertManager.postVectorAlert(
-                DefaultVectorAlert(
-                        uid = "enablePush",
-                        title = getString(R.string.alert_push_are_disabled_title),
-                        description = getString(R.string.alert_push_are_disabled_description),
-                        iconId = R.drawable.ic_room_actions_notifications_mutes,
-                        shouldBeDisplayedIn = {
-                            it is TchapHomeActivity
-                        }
-                ).apply {
-                    colorInt = ThemeUtils.getColor(this@TchapHomeActivity, R.attr.vctr_notice_secondary)
-                    contentAction = Runnable {
-                        (weakCurrentActivity?.get() as? VectorBaseActivity<*>)?.let {
-                            // action(it)
-                            homeActivityViewModel.handle(HomeActivityViewActions.PushPromptHasBeenReviewed)
-                            it.navigator.openSettings(it, VectorSettingsActivity.EXTRA_DIRECT_ACCESS_NOTIFICATIONS)
-                        }
-                    }
-                    dismissedAction = Runnable {
-                        homeActivityViewModel.handle(HomeActivityViewActions.PushPromptHasBeenReviewed)
-                    }
-                    addButton(getString(R.string.dismiss), {
-                        homeActivityViewModel.handle(HomeActivityViewActions.PushPromptHasBeenReviewed)
-                    }, true)
-                    addButton(getString(R.string.settings), {
-                        (weakCurrentActivity?.get() as? VectorBaseActivity<*>)?.let {
-                            // action(it)
-                            homeActivityViewModel.handle(HomeActivityViewActions.PushPromptHasBeenReviewed)
-                            it.navigator.openSettings(it, VectorSettingsActivity.EXTRA_DIRECT_ACCESS_NOTIFICATIONS)
-                        }
-                    }, true)
-                }
-        )
-    }
-
-    private fun promptSecurityEvent(userItem: MatrixItem.UserItem?, titleRes: Int, descRes: Int, action: ((VectorBaseActivity<*>) -> Unit)) {
-        popupAlertManager.postVectorAlert(
-                VerificationVectorAlert(
-                        uid = "upgradeSecurity",
-                        title = getString(titleRes),
-                        description = getString(descRes),
-                        iconId = R.drawable.ic_shield_warning
-                ).apply {
-                    viewBinder = VerificationVectorAlert.ViewBinder(userItem, avatarRenderer)
-                    colorInt = ContextCompat.getColor(this@TchapHomeActivity, R.color.riotx_positive_accent)
-                    contentAction = Runnable {
-                        (weakCurrentActivity?.get() as? VectorBaseActivity<*>)?.let {
-                            action(it)
-                        }
-                    }
-                    dismissedAction = Runnable {}
-                }
-        )
-    }
-
-    override fun onNewIntent(intent: Intent?) {
-        super.onNewIntent(intent)
-        if (intent?.getParcelableExtra<HomeActivityArgs>(MvRx.KEY_ARG)?.clearNotification == true) {
-            notificationDrawerManager.clearAllEvents()
-        }
-        handleIntent(intent)
-    }
-
-    override fun onDestroy() {
-        views.drawerLayout.removeDrawerListener(drawerListener)
-        super.onDestroy()
-    }
-
-    override fun onResume() {
-        super.onResume()
-
-        if (vectorUncaughtExceptionHandler.didAppCrash(this)) {
-            vectorUncaughtExceptionHandler.clearAppCrashStatus(this)
-
-            AlertDialog.Builder(this)
-                    .setMessage(R.string.send_bug_report_app_crashed)
-                    .setCancelable(false)
-                    .setPositiveButton(R.string.yes) { _, _ -> bugReporter.openBugReportScreen(this) }
-                    .setNegativeButton(R.string.no) { _, _ -> bugReporter.deleteCrashFile(this) }
-                    .show()
-        } else {
-            showDisclaimerDialog(this)
-        }
-
-        // Force remote backup state update to update the banner if needed
-        serverBackupStatusViewModel.refreshRemoteStateIfNeeded()
-    }
-
-    override fun configure(toolbar: Toolbar) {
-        configureToolbar(toolbar, false)
-    }
-
-    override fun getMenuRes() = R.menu.tchap_menu_home
-
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        when (item.itemId) {
-            R.id.menu_home_filter -> {
-                navigator.openRoomsFiltering(this)
-                return true
-            }
-        }
-
-        return super.onOptionsItemSelected(item)
-    }
-
-    override fun onBackPressed() {
-        if (views.drawerLayout.isDrawerOpen(GravityCompat.START)) {
-            views.drawerLayout.closeDrawer(GravityCompat.START)
-        } else {
-            super.onBackPressed()
-        }
-    }
-
-    override fun navToMemberProfile(userId: String, deepLink: Uri): Boolean {
-        val listener = object : MatrixToBottomSheet.InteractionListener {
-            override fun navigateToRoom(roomId: String) {
-                navigator.openRoom(this@TchapHomeActivity, roomId)
-            }
-        }
-        // TODO check if there is already one??
-        MatrixToBottomSheet.withLink(deepLink.toString(), listener)
-                .show(supportFragmentManager, "HA#MatrixToBottomSheet")
-        return true
-    }
-
-    companion object {
-        fun newIntent(context: Context, clearNotification: Boolean = false, accountCreation: Boolean = false): Intent {
-            val args = HomeActivityArgs(
-                    clearNotification = clearNotification,
-                    accountCreation = accountCreation
-            )
-
-            return Intent(context, TchapHomeActivity::class.java)
-                    .apply {
-                        putExtra(MvRx.KEY_ARG, args)
-                    }
-        }
-
-        private const val MATRIX_TO_CUSTOM_SCHEME_URL_BASE = "element://"
-        private const val ROOM_LINK_PREFIX = "${MATRIX_TO_CUSTOM_SCHEME_URL_BASE}room/"
-        private const val USER_LINK_PREFIX = "${MATRIX_TO_CUSTOM_SCHEME_URL_BASE}user/"
-    }
-}
