+++ vector/src/main/java/im/vector/app/core/di/FragmentModule.kt
@@ -22,8 +22,6 @@ import androidx.fragment.app.FragmentFactory
 import dagger.Binds
 import dagger.Module
 import dagger.multibindings.IntoMap
-import fr.gouv.tchap.features.home.TchapHomeDetailFragment
-import fr.gouv.tchap.features.home.room.list.TchapRoomListFragment
 import fr.gouv.tchap.features.login.TchapLoginFragment
 import fr.gouv.tchap.features.login.TchapWelcomeFragment
 import im.vector.app.features.attachments.preview.AttachmentsPreviewFragment
@@ -632,14 +630,4 @@ interface FragmentModule {
     @IntoMap
     @FragmentKey(TchapLoginFragment::class)
     fun bindTchapLoginFragment(fragment: TchapLoginFragment): Fragment
-
-    @Binds
-    @IntoMap
-    @FragmentKey(TchapHomeDetailFragment::class)
-    fun bindTchapHomeDetailFragment(fragment: TchapHomeDetailFragment): Fragment
-
-    @Binds
-    @IntoMap
-    @FragmentKey(TchapRoomListFragment::class)
-    fun bindTchapRoomListFragment(fragment: TchapRoomListFragment): Fragment
 }
