+++ vector/src/main/java/im/vector/app/core/di/ScreenComponent.kt
@@ -21,7 +21,6 @@ import androidx.fragment.app.FragmentFactory
 import androidx.lifecycle.ViewModelProvider
 import dagger.BindsInstance
 import dagger.Component
-import fr.gouv.tchap.features.home.TchapHomeActivity
 import fr.gouv.tchap.features.login.TchapLoginActivity
 import im.vector.app.core.dialogs.UnrecognizedCertificateDialog
 import im.vector.app.core.error.ErrorFormatter
@@ -154,7 +153,6 @@ interface ScreenComponent {
     fun inject(activity: ReAuthActivity)
     fun inject(activity: RoomDevToolActivity)
     fun inject(activity: TchapLoginActivity)
-    fun inject(activity: TchapHomeActivity)
 
     /* ==========================================================================================
      * BottomSheets
