+++ vector/src/main/java/im/vector/app/features/home/room/list/RoomListDisplayModeFilter.kt
@@ -31,7 +31,7 @@ class RoomListDisplayModeFilter(private val displayMode: RoomListDisplayMode) : 
             RoomListDisplayMode.NOTIFICATIONS ->
                 roomSummary.notificationCount > 0 || roomSummary.membership == Membership.INVITE || roomSummary.userDrafts.isNotEmpty()
             RoomListDisplayMode.PEOPLE        -> roomSummary.isDirect && roomSummary.membership.isActive()
-            RoomListDisplayMode.ROOMS         -> !roomSummary.isDirect && roomSummary.membership.isActive()
+            RoomListDisplayMode.ROOMS         -> roomSummary.membership.isActive()
             RoomListDisplayMode.FILTERED      -> roomSummary.membership == Membership.JOIN
         }
     }
